+++ curve25519/src/edwards.rs
@@ -9,7 +9,7 @@ pub struct PodEdwardsPoint(pub [u8; 32]);
 mod target_arch {
     use {
         super::*,
-        crate::curve25519::{
+        crate::{
             curve_syscall_traits::{GroupOperations, MultiScalarMultiplication, PointValidation},
             errors::Curve25519Error,
             scalar::PodScalar,
@@ -134,7 +134,7 @@ mod target_arch {
 mod target_arch {
     use {
         super::*,
-        crate::curve25519::{
+        crate::{
             curve_syscall_traits::{ADD, CURVE25519_EDWARDS, MUL, SUB},
             scalar::PodScalar,
         },
@@ -245,7 +245,7 @@ mod target_arch {
 mod tests {
     use {
         super::*,
-        crate::curve25519::scalar::PodScalar,
+        crate::scalar::PodScalar,
         curve25519_dalek::{
             constants::ED25519_BASEPOINT_POINT as G, edwards::EdwardsPoint, traits::Identity,
         },
