+++ curve25519/src/scalar.rs
@@ -6,7 +6,7 @@ pub struct PodScalar(pub [u8; 32]);
 
 #[cfg(not(target_os = "solana"))]
 mod target_arch {
-    use {super::*, crate::curve25519::errors::Curve25519Error, curve25519_dalek::scalar::Scalar};
+    use {super::*, crate::errors::{Curve25519Error, ElGamalError}, curve25519_dalek::scalar::Scalar};
 
     impl From<&Scalar> for PodScalar {
         fn from(scalar: &Scalar) -> Self {
@@ -19,6 +19,20 @@ mod target_arch {
 
         fn try_from(pod: &PodScalar) -> Result<Self, Self::Error> {
             Scalar::from_canonical_bytes(pod.0).ok_or(Curve25519Error::PodConversion)
+        }
+    }
+
+    impl From<Scalar> for PodScalar {
+        fn from(scalar: Scalar) -> Self {
+            Self(scalar.to_bytes())
+        }
+    }
+
+    impl TryFrom<PodScalar> for Scalar {
+        type Error = ElGamalError;
+
+        fn try_from(pod: PodScalar) -> Result<Self, Self::Error> {
+            Scalar::from_canonical_bytes(pod.0).ok_or(ElGamalError::CiphertextDeserialization)
         }
     }
 }
