--- zk-token-sdk/src/curve25519/errors.rs
@@ -1,7 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Clone, Debug, Eq, PartialEq)]
-pub enum Curve25519Error {
-    #[error("pod conversion failed")]
-    PodConversion,
-}
