+++ zk-token-sdk/src/encryption/elgamal.rs
@@ -21,7 +21,6 @@ use {
                 Pedersen, PedersenCommitment, PedersenOpening, G, H, PEDERSEN_COMMITMENT_LEN,
             },
         },
-        errors::ElGamalError,
         RISTRETTO_POINT_LEN, SCALAR_LEN,
     },
     base64::{prelude::BASE64_STANDARD, Engine},
@@ -40,6 +39,7 @@ use {
             SeedDerivable, Signer, SignerError,
         },
     },
+    solana_curve25519::errors::ElGamalError,
     std::convert::TryInto,
     subtle::{Choice, ConstantTimeEq},
     zeroize::Zeroize,
