+++ zk-token-sdk/src/zk_token_elgamal/pod/pedersen.rs
@@ -2,8 +2,9 @@
 
 #[cfg(not(target_os = "solana"))]
 use {
-    crate::{encryption::pedersen as decoded, errors::ElGamalError},
+    crate::encryption::pedersen as decoded,
     curve25519_dalek::ristretto::CompressedRistretto,
+    solana_curve25519::errors::ElGamalError,
 };
 use {
     crate::{
